Rewrite data URIs as blob URIs internally during stylesheet parsing
When parsing a stylesheet containing a large string the regex utilized by Dompdf can fail. Depending on the location of the problematic string, the regex failure can prevent parsing a style property, a style declaration, or an entire stylesheet.

This change is intended to resolve the most likely cause of this type of failure, which is the presence of data URIs in the stylesheet. To work around the regex parsing issues Dompdf replaces data URIs with blob-like URIs. The blob URI shortens the string to a regex-parsable length. When the CSS property value is retrieved the blob URI is rewritten to it's original form.

fixes #3482