Update Cpdf font metrics handling
MissingWidth: Some monospace fonts do not include the MissingWidth value in the font metrics (see DejaVu Sans Mono). In this scenario use the width of the space character if it exists.

CapHeight: FontLib now retrieves the CapHeight, so use that value if present. Otherwise continue to use the Ascender value.