Support custom property references in psuedo-element content
Prior to this change a pseudo-element would not be generated if it's only content was a custom property reference. This is because the frame is generated before styles are applied and the custom property reference is unable to resolve. The lack of content in a pseudo-element precludes frame generation.

Even if a frame is generated, it is not rendered to the page if it has no content. This change modifies the content check to exclude only if there is no content (null) rather than empty ("").

fixes #2775 and #3573